requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-deepgram
livekit-plugins-silero
livekit-plugins-cartesia
livekit-plugins-azure
python-dotenv
requests
Flask
livekit-api